Add clear loop optimization, better documentation and debug output.
 - Clear loops ([-] and [+]) are now optimized out. This actually
   helps performance a lot, especially in the alphabet example, and
   is rather simple with the way we parse these.
 - DEBUG option for Makefile
 - Some more documentation in comments
 - Warning flags in Makefile
 - Updated README.md with info about implementation and conformance
 - Intro and cleanup added in DEBUG output
 - Used printf reliably for consistency in DEBUG output
 - Use spaces instead of tabs for DEBUG output. This improves readability
   in the source file.

+++ main.c
@@ -23,7 +23,7 @@ int main(int argc, char *argv[])
             close(fd);
             return 1;
         }
-        size_t len = st.st_size;
+        off_t len = st.st_size;
         char *buf = (char *)malloc(len + 1);
         if (!buf) {
             puts("Out of memory");
